mikrotik/pyproject.toml
[project]
name = "mikrotik"
version = "1.0.0"
description = "Helper scripts for Mikrotik setup"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]
